Create MazeRunner.java
This is the Maze Runner class. The class allows the "Maze Runner" object to get the user's move, ensure that the user's move is valid (for example, there should not be a wall), and navigate through pits. Additionally, a message will also be printed depending on the number of moves left.