Add files via upload
This is the first draft of my MazeRunner class. It creates the MazeRunner object, has the introduction method, the game method (which checks to see if the user has successfully escaped the maze), gets the user's move (with the additional method "getValidStringInput"), and checks if the move is valid before moving.